modifying

+++ binary-tarball-tests/pxc/run.sh
@@ -4,9 +4,10 @@ export PATH=${HOME}/.local/bin:${PATH}
 PXC_MAJOR_VERSION="$(echo ${PXC_VERSION}|cut -d'.' -f1,2)"
 
 echo "Installing dependencies..."
-if grep -q 'Amazon Linux' /etc/os-release && grep -q '2023' /etc/os-release; then
+if [ -f /etc/os-release ] && grep -q 'Amazon Linux' /etc/os-release && grep -q '2023' /etc/os-release; then
   echo "Detected Amazon Linux 2023"
   sudo yum install -y python3 python3-pip libaio numactl openssl socat lsof libev
+  export PATH="${HOME}/.local/bin:${PATH}"
   pip3 install --user pytest-testinfra pytest
 elif [ -f /etc/redhat-release ]; then
   sudo yum install -y libaio numactl openssl socat lsof
